+++ auth0_flutter/test/web/auth0_extension_type_mocks.dart
@@ -23,3 +23,8 @@ class WebCredentialsImpl {
   @JSExport('access_token')
   String accessToken = throw Exception('');
 }
+
+@JSExport()
+class PasskeyChallengeResponseImpl {
+  String authSession = throw Exception('');
+}